Some refactoring. Heroku adaptation and sysconfig vars
