+++ .gitignore
@@ -1,0 +1,6 @@
+pip-selfcheck.json
+ds_mailer/creds.py
+__pycache__/
+.vscode/
+.idea/
+venv/
