--- ds_mailer/.gitignore
@@ -1,6 +1,0 @@
-Include/
-Lib/
-pip-selfcheck.json
-Scripts/
-tcl/
-mailer/creds.py
