+++ ds_mailer/tests/test_gamelist.py
@@ -1,0 +1,20 @@
+from ds_mailer.Game import Game
+
+
+TEST_GAMES = [
+    Game("DARK SOULS 1. PS4 Remastered",
+         "https://store.playstation.com/ru-ru/product/EP0700-CUSA08495_00-DARKSOULSHD00000",
+         "2.699"),
+    Game("DARK SOULS 2. PS4 Scholar of the first sin",
+         "https://store.playstation.com/ru-ru/product/EP0700-CUSA01589_00-DARKSOULS2000003",
+         "1.199"),
+    Game("DARK SOULS 3. PS4 Delux Edition",
+         "https://store.playstation.com/ru-ru/product/EP0700-CUSA03365_00-DSIIIDELUXE00000",
+         "4.999"),
+    Game("Bloodbourne. PS4 Game of the year",
+         "https://store.playstation.com/ru-ru/product/EP9000-CUSA03173_00-BLOODBORNE0000EU",
+         "1.999"),
+    Game("Last of us 2",
+         "https://store.playstation.com/ru-ru/product/EP9000-CUSA10249_00-THELASTOFUSPART2",
+         "5.499")
+]
