+++ ds_mailer/tests/test_sender.py
@@ -1,0 +1,6 @@
+from ds_mailer.Sender import run_sender
+from ds_mailer.tests.test_gamelist import TEST_GAMES
+
+
+if __name__ == "__main__":
+    run_sender(TEST_GAMES)
